refactor: languages page with components
references #344
